expand operator with a passed testcase
